Moved classes to library
