+++ cobotSimulatorTwo/pom.xml
@@ -23,7 +23,7 @@
             <artifactId>org.eclipse.paho.client.mqttv3</artifactId>
             <version>1.2.5</version>
         </dependency>
-        <dependency> <!--Doesn't recognize the package for some reason-->
+        <dependency>
             <groupId>javiergs</groupId>
             <artifactId>library</artifactId>
             <version>1.0-SNAPSHOT</version>
