+++ cobotSimulatorTwo/src/main/java/cobot/ProgressBar.java
@@ -1,5 +1,7 @@
 package cobot;
 
+import CobotSimulatorTwo.Blackboard;
+
 import javax.swing.*;
 import java.awt.*;
 import java.beans.PropertyChangeEvent;
