+++ cobotSimulatorTwo/src/main/java/cobot/RobotPanelHandler.java
@@ -1,5 +1,7 @@
 package cobot;
 
+import CobotSimulatorTwo.Blackboard;
+
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
