+++ library/src/main/java/CobotSimulatorTwo/Subscriber.java
@@ -1,4 +1,4 @@
-package cobot;
+package CobotSimulatorTwo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
